+++ packages/components/atoms/f-link/src/tenants/en-GB.js
@@ -1,10 +1,6 @@
-const messages = {
+export default {
     ariaDescription: {
         newLocation: 'Opens in a new window/screen/tab',
-        externalSite: 'Opens an external site in a new window/screen/tab'
+        external: 'Opens an external site in a new window/screen/tab'
     }
-};
-
-export default {
-    messages
 };