+++ packages/components/atoms/f-link/src/tenants/es-ES.js
@@ -1,10 +1,6 @@
-const messages = {
+export default {
     ariaDescription: {
         newLocation: 'Abrir una nueva ventana/pestaña',
-        externalSite: 'Abrir a una web externa en una nueva ventana/pestaña'
+        external: 'Abrir a una web externa en una nueva ventana/pestaña'
     }
-};
-
-export default {
-    messages
 };