+++ packages/components/atoms/f-link/src/tenants/it-IT.js
@@ -1,10 +1,6 @@
-const messages = {
+export default {
     ariaDescription: {
         newLocation: 'Stai aprendo una nuova finestra/schermata/scheda',
-        externalSite: 'Stai aprendo un sito esterno in una nuova finestra/schermata/scheda'
+        external: 'Stai aprendo un sito esterno in una nuova finestra/schermata/scheda'
     }
-};
-
-export default {
-    messages
 };