+++ packages/components/atoms/f-link/stories/Link.stories.js
@@ -3,7 +3,7 @@ import {
 } from '@storybook/addon-knobs';
 import { withA11y } from '@storybook/addon-a11y';
 import VLink from '../src/components/Link.vue';
-
+import { DEFAULT_LINK_TYPE, VALID_LINK_TYPE, VALID_LINK_TYPES }  from '../src/constants'
 export default {
     title: 'Components/Atoms',
     decorators: [withA11y]
@@ -17,12 +17,8 @@ export const VLinkComponent = () => ({
         }
     },
     props: {
-        isExternal: {
-            default: boolean('isExternal', false)
-        },
-
-        opensInNewLocation: {
-            default: boolean('opensInNewLocation', false)
+        linkType: {
+            default: select('Link Type', VALID_LINK_TYPES, DEFAULT_LINK_TYPE)
         },
 
         isBold: {
@@ -45,16 +41,13 @@ export const VLinkComponent = () => ({
     template: `<v-link
                     :data-test-id="dataTestId"
                     href="https://www.just-eat.co.uk/"
-                    :is-external="isExternal"
-                    :opens-in-new-location="opensInNewLocation"
+                    :link-type="linkType"
                     :is-bold="isBold"
                     :has-text-decoration="hasTextDecoration"
                     :is-full-width="isFullWidth"
                     :no-line-break="noLineBreak"
                 >
-                    <span>
-                        This is a Link
-                    </span>
+                    <span>This is a Link</span>
                 </v-link>`
 });