+++ packages/components/atoms/f-link/CHANGELOG.md
@@ -4,6 +4,18 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+v0.3.0
+------------------------------
+*May 26, 2021*
+
+### Removed
+- `VueGlobalisationMixin`
+
+### Changed
+- `isExternalLink` and `opensInNewLocation` props to 'linkType'
+- Tests to cover changes.
+
+
 v0.2.0
 ------------------------------
 *May 24, 2021*